requirements.txt
# --- Core packages ---
streamlit>=1.35.0
moviepy>=1.0.3
numpy>=1.22.0
pandas>=1.4.0
Pillow>=10.0.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0.76
python-dotenv>=1.0.0

# --- OCR & TTS ---
pytesseract>=0.3.10
gTTS>=2.3.2

# --- Document processing ---
python-docx>=1.1.0
openpyxl>=3.1.2
pypdf>=3.17.4

# --- LLM & Embeddings ---
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
langchain-text-splitters>=0.0.1

# --- Streamlit components ---
streamlit-ace>=0.1.1
streamlit-plotly-events>=0.0.6
streamlit-mic-recorder>=0.0.6

# --- Utilities ---
requests>=2.31.0
beautifulsoup4>=4.12.3
tenacity>=8.2.3

streamlit
groq
python-dotenv
streamlit-mic-recorder
gTTS